Feature backend permissions

+++ backend/geanihouse/src/main/java/ies/proj/geanihouse/controller/DeviceController.java
@@ -4,10 +4,12 @@ import ies.proj.geanihouse.model.Division;
 import ies.proj.geanihouse.model.Device;
 import ies.proj.geanihouse.repository.DeviceRepository;
 import ies.proj.geanihouse.repository.DivisionRepository;
+import ies.proj.geanihouse.service.PermissionService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -45,12 +47,20 @@ public class DeviceController {
     @Autowired
     Source source;
 
-    @GetMapping("/{id}/devices/")
-    public ResponseEntity<?> getAllHomeDivisions(@PathVariable(value = "id") Long id) throws ResourceNotFoundException {
-        Division division = this.divisionRepository.findById(id).orElseThrow( () -> new ResourceNotFoundException("Could not find division with id" + id));
-        Set <Device> devices = division.getDevices();
+    @Autowired
+    private PermissionService permissionService;
 
-        System.out.println(devices);
+    private UserDetails authenticateduser;
+
+    @GetMapping("/{id}/devices/")
+    public ResponseEntity<?> getAllDivisionDevices(@PathVariable(value = "id") Long id) throws ResourceNotFoundException {
+        Division division = this.divisionRepository.findById(id).
+                orElseThrow( () -> new ResourceNotFoundException("Could not find division with id" + id));
+        this.authenticateduser= (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!this.permissionService.checkClientDivision(division,this.authenticateduser)){
+           return  ResponseEntity.status(403).body("Cannot get Devices from a House you Dont Belong!");
+        }
+        Set <Device> devices = division.getDevices();
 
         return ResponseEntity.ok().body(devices);
     }
@@ -59,6 +69,11 @@ public class DeviceController {
     public ResponseEntity<?> addDeviceToDivision(@Valid @RequestBody Device device) throws ResourceNotFoundException {
         Division d = divisionRepository.findById(device.getDivision().getId())
         .orElseThrow(() -> new ResourceNotFoundException("Could not find division "));
+
+        this.authenticateduser= (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!this.permissionService.checkClientDivision(d,this.authenticateduser)){
+            return  ResponseEntity.status(403).body("Cannot add Devices to a House you Dont Belong!");
+        }
 
         Type t = typeRepository.findById(device.getType().getId()).
         orElseThrow(() -> new ResourceNotFoundException("Could not find Type of Sensor "));
@@ -82,7 +97,11 @@ public class DeviceController {
     @PutMapping("/devices")
     public ResponseEntity<?> updateDevice(@Valid @RequestBody Device device) throws ResourceNotFoundException {
         Device n = deviceRepository.findById(device.getId()).
-        orElseThrow(() -> new ResourceNotFoundException("Could not find Type of Sensor "));;
+        orElseThrow(() -> new ResourceNotFoundException("Could not find Type of Sensor "));
+        this.authenticateduser= (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!this.permissionService.checkClientDivision(n.getDivision(),this.authenticateduser)){
+            ResponseEntity.status(403).body("Cannot update a Device from a House you Dont Belong!");
+        }
         n.setState(device.getState());
         deviceRepository.save(n);
         return  ResponseEntity.ok().body("Successfully added new Device");
@@ -93,9 +112,15 @@ public class DeviceController {
     public Map<String,Boolean> deleteDevice(@PathVariable(value = "id") Long deviceId) throws ResourceNotFoundException {
         Device device = deviceRepository.findById(deviceId)
                 .orElseThrow( () -> new ResourceNotFoundException("Division not found for this id :: " + deviceId));
-        LOG.debug("Deleteting device " + device);
-        deviceRepository.delete(device);
         Map<String,Boolean> response = new HashMap<>();
+        if(! permissionService.checkClientDivision(device.getDivision(),this.authenticateduser)){
+            // Forbidden!
+            response.put("deleted",Boolean.FALSE);
+            return response;
+        }
+        LOG.info("Deleting device " + device);
+        deviceRepository.delete(device);
+
         response.put("deleted",Boolean.TRUE);
         return response;
     }
