+++ backend/geanihouse/src/main/java/ies/proj/geanihouse/controller/DivisionConfController.java
@@ -10,10 +10,13 @@ import ies.proj.geanihouse.model.Type;
 import ies.proj.geanihouse.repository.DivisionConfRepository;
 import ies.proj.geanihouse.repository.DivisionRepository;
 import ies.proj.geanihouse.repository.TypeRepository;
+import ies.proj.geanihouse.service.PermissionService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
@@ -37,6 +40,8 @@ public class DivisionConfController {
     @Autowired
     private TypeRepository typeRepository;
 
+    @Autowired
+    private PermissionService permissionService;
 
     @GetMapping("/divisions/configurations")
     public List<DivisionConf> getAllDivisionConfs(){
@@ -44,21 +49,27 @@ public class DivisionConfController {
     }
     @GetMapping("/divisions/{id}/configurations")
     public ResponseEntity<?> getDivisionConfs(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
-
         Division division = divisionRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException ("Could not find division with id :: " + id));
+        UserDetails authenticateduser = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!this.permissionService.checkClientDivision(division, authenticateduser)){
+            return  ResponseEntity.status(403).body("Cannot get Configurations of a Division you Dont Belong!");
+        }
         LOG.info("Returning division Configurations for division with id  " + id);
         List<DivisionConf> divisionConfs = divisionConfRepository.findAllByDivisionId(id);
         return  ResponseEntity.ok().body(divisionConfs);
-
     }
 
     @PostMapping("/divisions/configurations")
-    public DivisionConf addDivisionConfs(@Valid @RequestBody DivisionConf divisionConf) throws  ResourceNotFoundException, ErrorDetails{
+    public ResponseEntity<?> addDivisionConfs(@Valid @RequestBody DivisionConf divisionConf) throws  ResourceNotFoundException, ErrorDetails{
 
         //check if division exists
         Division division= divisionRepository.findById(divisionConf.getDivision().getId())
                 .orElseThrow(() -> new ResourceNotFoundException("Could not Find division with id :: "+ divisionConf.getDivision().getId()));
 
+        UserDetails authenticateduser = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!this.permissionService.checkClientDivision(division, authenticateduser)){
+            return  ResponseEntity.status(403).body("Cannot add Configurations of a Division you Dont Belong!");
+        }
         //check if the type is valid
 
         Type type = typeRepository.findByName(divisionConf.getType().getName());
@@ -68,7 +79,6 @@ public class DivisionConfController {
         }
         divisionConf.setType(type);
 
-
         //check the contraints to create a new Configuration
         String error_message = checkValuesContraints(divisionConf);
 
@@ -77,7 +87,8 @@ public class DivisionConfController {
         if(error_message != null)throw new ErrorDetails(error_message);
 
         LOG.info("Success inserting new Configuration for division " + division.getId());
-        return divisionConfRepository.save(divisionConf);
+        divisionConfRepository.save(divisionConf);
+        return ResponseEntity.ok().body("Success inserting new Configuration");
     }
 
 
@@ -85,7 +96,12 @@ public class DivisionConfController {
     public ResponseEntity<?> updateDivisionConf(@PathVariable(value = "id") long id, @Valid @RequestBody DivisionConf dconf) throws ResourceNotFoundException, ErrorDetails{
         DivisionConf divisionConf = divisionConfRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Could not find division configuration with id ::" + id));
-        
+
+
+        UserDetails authenticateduser = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!this.permissionService.checkClientDivision(divisionConf.getDivision(), authenticateduser)){
+            return  ResponseEntity.status(403).body("Cannot update Configurations of a Division you Dont Belong!");
+        }
         LOG.info("PUT Request DivisionConf");
         // verify if the type exists
         Type type = typeRepository.findByName(divisionConf.getType().getName());
@@ -109,8 +125,13 @@ public class DivisionConfController {
     public Map<String,Boolean> deleteDivisionConf(@PathVariable(value ="id") long id) throws ResourceNotFoundException{
         DivisionConf divisionConf = divisionConfRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Could not find division configuration with id ::" + id));
-        divisionConfRepository.delete(divisionConf);
+        UserDetails authenticateduser = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Map <String,Boolean> response = new HashMap<>();
+        if(!this.permissionService.checkClientDivision(divisionConf.getDivision(), authenticateduser)){
+            response.put("deleted",Boolean.FALSE);
+            return  response;
+        }
+        divisionConfRepository.delete(divisionConf);
         response.put("deleted",Boolean.TRUE);
         return response;
     }
